backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Document conversion
python-docx==1.1.0
pdf2docx==0.5.8
reportlab==4.1.0
openpyxl==3.1.4

# Image processing
Pillow==10.3.0

# PDF tools
PyMuPDF==1.24.3
